Refactor AnySigner and remove unnecessary C interfaces (#862)
* remove AnySigner will be replaced by TrustSDK
* Add missing TWVechainSignerTests
* Add AnySigner (c++ and swift)
* Handle Filecoin and FIO
* AnySigner for Android
* Remove more exposed C interfaces
*  Move signer tests to blockchain folder
*  Fix potensial swift memory leak
* Add NEO to AnySigner
* Bitcoin AnySigner
* add Decred AnySigner tests

+++ src/Bitcoin/UnspentCalculator.cpp
@@ -31,7 +31,7 @@ UnspentCalculator UnspentCalculator::getCalculator(TWCoinType coinType) {
     }
 }
 
-int64_t UnspentCalculator::calculateFee(size_t inputs, size_t outputs, int64_t byteFee) {
+int64_t UnspentCalculator::calculateFee(int64_t inputs, int64_t outputs, int64_t byteFee) {
     const auto txsize = ((148 * inputs) + (34 * outputs) + 10);
     return int64_t(txsize) * byteFee;
 }
